anomaly-detection-container/requirements.txt
kafka-python
psycopg2-binary
influxdb
numpy
kagglehub